Clean up doc pages and consolidate base classes.
