+++ doc/conf.py
@@ -125,6 +125,7 @@ html_static_path = ['_static']
 # https://github.com/rtfd/sphinx_rtd_theme/issues/117
 def setup(app):
     app.add_stylesheet('theme_overrides.css')
+    app.connect('autodoc-process-docstring', generate_example_rst)
 
 
 html_favicon = '_static/nimare_favicon.png'
@@ -186,10 +187,6 @@ def generate_example_rst(app, what, name, obj, options, lines):
     if not os.path.exists(examples_path):
         # touch file
         open(examples_path, 'w').close()
-
-
-def setup(app):
-    app.connect('autodoc-process-docstring', generate_example_rst)
 
 # Documents to append as an appendix to all manuals.
 #texinfo_appendices = []
