+++ nimare/base/decode.py
