+++ nimare/dataset/__init__.py
@@ -1,7 +1,6 @@
 """
 Top-level namespace for datasets.
 """
-from .base import Contrast
 from .data import Database, Dataset
 from .extract import (NeuroVaultDataSource, NeurosynthDataSource,
                       BrainSpellDataSource)
