+++ nimare/dataset/extract.py
@@ -9,7 +9,7 @@ from os.path import join, isdir, isfile
 import pandas as pd
 from pyneurovault import api
 
-from .base import DataSource
+from ..base import DataSource
 from ..utils import get_resource_path
